requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
torch>=2.0.0
pyyaml>=6.0
mlflow>=2.8.0
matplotlib>=3.7.0
tqdm>=4.65.0
